refactor AddApplicationInsightsSettings - test pass
